Create Tries: Contacts
We're going to make our own Contacts application! The application must perform two types of operations:
add name, where  is a string denoting a contact name. This must store  as a new contact in the application.
find partial, where  is a string denoting a partial name to search the application for. It must count the number of contacts starting with  and print the count on a new line.